Updating documentations
