+++ Sources/Guaka/Flag/Flag.swift
@@ -7,34 +7,76 @@
 //
 
 
-/// Flag struct
+/// Flag struct representing a flag
+///
+/// ----
+/// Example
+/// Create a flag with a default value
+/// ```
+/// let flag = Flag(longName: "debug", shortName: "d", value: true, description: "Here is a desc")
+/// ```
+///
+/// Create a non required flag without value
+/// ```
+/// let flag = Flag(longName: "debug", type: Int.self)
+/// ```
+///
+/// Create a required flag without value
+/// ```
+/// let flag = Flag(longName: "debug", type: Int.self, required: true)
+/// ```
 public struct Flag: Hashable {
 
-  /// Flag long name
+
+  /// Flag long name. Such as `verbose`
   public let longName: String
 
-  /// Flag short name
+
+  /// Flag short name. Should be one letter. Such as `v`
   public let shortName: String?
+
 
   /// Set the flag to be inheritable
   /// An inheritable flag is valid for a command and all its subcommands
   public let inheritable: Bool
 
-  /// The flag description
+
+  /// The flag description printed in the help beside the flag name
   public let description: String
 
-  /// Flag value type
-  public let type: CommandStringConvertible.Type
 
-  /// Set the flag to be required
-  // A required flag must be set in order for commands to be executed
+  /// Flag value type.
+  public let type: FlagValueStringConvertible.Type
+
+
+  /// Set the flag to be required/
+  /// A required flag must be set in order for commands to be executed.
   public let required: Bool
 
-  /// Flag value
-  public var value: CommandStringConvertible?
 
-  /// Flag deprecation status
+  /// Flag value.
+  /// The value can be `int`, `bool`, `string` or any other type that implements `FlagValueStringConvertible`
+  public var value: FlagValueStringConvertible?
+
+
+  /// Flag deprecation status.
   public var deprecationStatus = DeprecationStatus.notDeprecated
+
+
+  /// Gets the flag hash value.
+  public var hashValue: Int {
+    return longName.hashValue
+  }
+
+
+  /// Return if the flag is a boolean or not.
+  var isBool: Bool {
+    return value is Bool
+  }
+
+
+  /// Was the flag value set from the arguments
+  var didSet: Bool = false
 
 
   /// Creats a new flag
@@ -46,7 +88,7 @@ public struct Flag: Hashable {
   /// - parameter description: Flag description to be shown when displaying command help
   public init(longName: String,
               shortName: String? = nil,
-              value: CommandStringConvertible,
+              value: FlagValueStringConvertible,
               inheritable: Bool = true,
               description: String = "") {
 
@@ -59,6 +101,7 @@ public struct Flag: Hashable {
     self.required = true
   }
 
+
   /// Creats a new flag
   ///
   /// - parameter longName:    Flag long name
@@ -69,7 +112,7 @@ public struct Flag: Hashable {
   /// - parameter description: Flag description to be shown when displaying command help
   public init(longName: String,
               shortName: String? = nil,
-              type: CommandStringConvertible.Type,
+              type: FlagValueStringConvertible.Type,
               required: Bool = false,
               inheritable: Bool = false,
               description: String = "") {
@@ -82,17 +125,10 @@ public struct Flag: Hashable {
     self.required = required
   }
 
-  public var hashValue: Int {
-    return longName.hashValue
-  }
-
-  var isBool: Bool {
-    return value is Bool
-  }
-
-  var didSet: Bool = false
 }
 
+
+/// Check for flags equality
 public func ==(left: Flag, right: Flag) -> Bool {
   return left.hashValue == right.hashValue
 }
