+++ Sources/Guaka/Misc/GuakaConfig.swift
@@ -6,6 +6,11 @@
 //
 //
 
+
+/// Guaka global configurations
 public struct GuakaConfig {
+
+  /// Sets the help generator strategy
+  /// By default this is `DefaultHelpGenerator`
   public static var helpGenerator: HelpGenerator.Type = DefaultHelpGenerator.self
 }
